Fix compiler warnings and pointer truncation for 64-bit builds.
Lots of casting to quiet MSVC 64-bit compiler warnings.
Fixed pointer truncation in TreeAlloc_Free().
Added extra bit of debug checking to TreeAlloc code.

+++ generic/qebind.c
@@ -230,7 +230,7 @@ int QE_InstallEvent(QE_BindingTable bindingTable, char *name, QE_ExpandProc expa
 	type = bindPtr->nextEventId++;
 
 	eiPtr = (EventInfo *) Tcl_Alloc(sizeof(EventInfo));
-	eiPtr->name = Tcl_Alloc(strlen(name) + 1);
+	eiPtr->name = Tcl_Alloc((int) strlen(name) + 1);
 	strcpy(eiPtr->name, name);
 	eiPtr->type = type;
 	eiPtr->expandProc = expandProc;
@@ -686,7 +686,7 @@ int QE_CreateBinding(QE_BindingTable bindingTable, ClientData object,
 	/* Append given command to any existing command */
 	if (append && cmdOld)
 	{
-		length = strlen(cmdOld) + strlen(command) + 2;
+		length = (int) (strlen(cmdOld) + strlen(command) + 2);
 		cmdNew = Tcl_Alloc((unsigned) length);
 		(void) sprintf(cmdNew, "%s\n%s", cmdOld, command);
 	}
@@ -1039,7 +1039,7 @@ static void ExpandPercents(BindingTable *bindPtr, ClientData object,
 		}
 		if (string != command)
 		{
-			Tcl_DStringAppend(result, command, string - command);
+			Tcl_DStringAppend(result, command, (int) (string - command));
 			command = string;
 		}
 		if (*command == 0)
